Update ssllabs.py
Update errors to statusMessage per api definition:

* **statusMessage** - status message in English. When status is ERROR, this field will contain an error message.

and errors is for incorrect API invoking:

When an API call is incorrectly invoked, it will cause an error response to be sent back. The response will include an array of error messages. For example:

    {"errors":[{"field":"host","message":"qp.mandatory"}]}

The field value references the API parameter name that has an incorrect value. The message value will tell you what the issue is. It is also possible to receive errors without the field parameter set; such messages will usually refer to the request as a whole.